Cast recovered to uint8 for imwrite()

+++ lecture_11/lecture_11_743.m
@@ -18,6 +18,6 @@ disp("recovering");
 recovered = restore_image(blurred, 100, 2000, 0.02, 3);
 
 disp("Saving recovered.png");
-imwrite(recovered, "recovered.png");
+imwrite(uint8(recovered), "recovered.png");
 
 disp("Done")
